Tidy

+++ SpatialGDK/Source/SpatialGDK/Private/LoadBalancing/VoronoiGridBasedLBStrategy.cpp
@@ -20,8 +20,8 @@ void UVoronoiGridBasedLBStrategy::InitVirtualWorkerIdGrid()
 {
 	const static TArray<FVector2D> Points = { FVector2D(-5000.f, -5000.f), FVector2D(5000.f, 5000.f), FVector2D(5000.f, -5000.f) , FVector2D(0.f, 0.f) };
 
-	const size_t Size = Points.Num();
-	for (uint32 i = 0; i < Size; ++i)
+	const int32 Size = Points.Num();
+	for (int32 i = 0; i < Size; ++i)
 	{
 		const FVector2D& Position = Points[i];
 		VirtualWorkerPositions.Add(i + 1, Position);
